Add link at the top of release-notes to the x.y.0 release notes
When a site wants to upgrade, it is likely that they do so targeting some
non-zero bug-fix version (i.e., not upgrade dCache to version x.y.0).
Therefore, it is possible (even likely) that the admin will not know to
navigate to the end of the page in order to see potentially critical
information.

To avoid major restructing of the release-notes, this patch adds a simple
link so the admin can discover the available information.